Added extra enemy, new background image and small text description fixes.

+++ src/Game/Character/Skill/Skills/MultiAttack.java
@@ -128,6 +128,6 @@ public class MultiAttack extends Skill {
     {
         super.setDescription("Do a sequence of " + this.getMaxAttacks() + 
                              " attacks that hit for " + this.damagePerHit + " damage each, with " + 
-                             (this.hitChance * PERCENTAGE) + "% chance to hit each one.");
+                             (int)(this.hitChance * PERCENTAGE) + "% chance to hit each one.");
     }
 }
